requirements.txt
pybaseball>=2.2.7
mlb-statsapi>=1.7.0
xgboost>=2.0.0
scikit-learn>=1.4.0
streamlit>=1.32.0
pandas>=2.2.0
numpy>=1.26.0
joblib>=1.3.0
plotly>=5.18.0
tqdm>=4.66.0
requests>=2.31.0
